Extract egg_link* function to a separate module
